deleted RpcEventFilter and filer_event, refactor get_event
